requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.5.1
ffmpeg-python==0.2.0
future==0.18.2
idna==2.10
multidict==4.7.6
typing-extensions==3.7.4.3
yarl==1.5.1
youtube-dl==2020.12.9